pyproject.toml
[project]
name = "ayne"
version = "0.1.0"
description = "End-to-end machine learning pipeline for movie & TV performance analysis and box-office prediction. Includes data ingestion, preprocessing, model training, experiment tracking, and simple web delivery."
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11,<3.13"
keywords = [
    "machine-learning",
    "data-science",
    "movies",
    "box-office",
    "ml-pipeline",
    "prediction",
    "mlflow"
]
authors = [
    {name = "Kristiyan Bonev", email = "k.s.bonev@gmail.com"}
]
maintainers = [
    {name = "Kristiyan Bonev", email = "k.s.bonev@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
    "Operating System :: OS Independent",
]
dependencies = [
    # --- Core scientific stack ---
    "numpy>=2.1,<3.0",
    "pandas>=2.2,<3.0",
    "polars>=1.7,<2.0", # Optional modern alternative – blazing fast
    "scikit-learn>=1.5,<2.0",
    "xgboost>=2.0,<3.0", # v2+ highly stable now
    # --- Visualization ---
    "matplotlib>=3.8,<4.0",
    "seaborn>=0.13,<0.15",
    "plotly>=6.0,<7.0", # Loose enough but avoids future breakage
    # --- Configuration / Serialization ---
    "pydantic>=2.0,<3.0", # Modern settings management
    "pydantic-settings>=2.0,<3.0", # Pydantic settings from env/files
    "python-dotenv>=1.0,<2.0",
    "pyyaml>=6.0,<7.0", # YAML parsing for config files
    "cloudpickle>=3.0,<4.0", # Better than dill for model serialization
    # --- Web / API ---
    "fastapi>=0.110,<1.0",
    "uvicorn>=0.30,<1.0",
    "httpx>=0.27,<1.0", # Modern async HTTP client
    # --- Database / ORM ---
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    # --- Utilities ---
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=5.2,<6.0", # Fast XML/HTML backend
    "pyprojroot>=0.3,<1.0", # Keep if you're using it
    # --- Experiment tracking ---
    "mlflow>=2.15,<2.20",
    "pyarrow>=16.1,<17.0",
    "duckdb>=1.4.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.6.0,<1.0.0",
    "mypy>=1.13.0,<2.0.0",
    "pip-audit>=2.9.0,<3.0.0",
    "bandit>=1.8.0,<2.0.0",
    "pymarkdownlnt>=0.9.17,<1.0.0"
]

llm-openai = [
    "openai>=1.0,<2.0"
]

llm-anthropic = [
    "anthropic>=0.40,<1.0"
]

notebooks = [
    "jupyterlab>=4.2,<5.0",
    "ipykernel>=6.29,<7.0",
    "ipywidgets>=8.1,<9.0",
    "widgetsnbextension>=4.0,<5.0",
    "matplotlib>=3.8,<4.0",
    "seaborn>=0.13,<0.15",
    "plotly>=6.0,<7.0",
    "wordcloud>=1.9,<2.0"
]

docs = [
    "mkdocs>=1.6.0,<2.0.0",
    "mkdocs-material>=9.5.0,<10.0.0",
    "mkdocstrings[python]>=0.27.0,<1.0.0",
    "pymdown-extensions>=10.0.0,<11.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ayne*"]
namespaces = false

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | notebooks
  | __pycache__
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
fix = true
exclude = [
    ".git",
    ".tox",
    "dist",
    "build",
    "*.egg",
    "reports",
    ".venv",
    "venv",
    "notebooks",
    ".pytest_cache",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "D"]  # Includes docstring checks
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in argument defaults
    "B904",   # Exception chaining (use `from` for clarity)
    "D203",   # 1 blank line before class (conflicts with D211)
    "D213",   # Multi-line docstring summary second line (conflicts with D212)
    "D205",   # 1 blank line between summary and description (too strict)
    "D415",   # First line punctuation (docstrings can be phrases)
    "D102",   # Missing docstring in public method (too verbose for simple methods)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
strict_optional = false
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
# Test discovery patterns
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test paths
testpaths = ["tests"]

# Additional options
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:reports/coverage/html",
    "--cov-report=xml:reports/coverage/coverage.xml",
    "--cov-fail-under=10",
    # "--strict-markers",
    # "--disable-warnings",
]

# Markers for categorizing tests
markers = [
    "unit: Unit tests for individual functions",
    "integration: Integration tests for API endpoints",
    "slow: Tests that take longer to run",
    "requires_model: Tests that require ML model loading",
]

# Ignore directories
norecursedirs = [
    ".git",
    ".tox",
    "dist",
    "build",
    "*.egg",
    "reports",
    ".venv",
    "venv",
    ".pytest_cache",
    "__pycache__",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/notebooks/*",
]
data_file = "reports/coverage/.coverage"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "reports/coverage/html"

[tool.coverage.xml]
output = "reports/coverage/coverage.xml"

[tool.bandit]
exclude_dirs = [
    ".venv",
    "venv",
    "tests",
    "notebooks",
    ".pytest_cache",
    "__pycache__",
]
skips = ["B104", "B108"]

# Markdown linting
[tool.pymarkdown]
plugins.md013.enabled = false  # Disable line length check (tables/code blocks can be long)
plugins.md033.enabled = false  # Allow inline HTML (useful for badges, images)
plugins.md013.line_length = 120  # Set reasonable line length if enabled
plugins.md036.enabled = false  # Allow emphasis as pseudo-headings (common in docs)
plugins.md024.enabled = false  # Allow duplicate headings (useful for examples)
plugins.md040.enabled = false  # Don't require language for all code blocks (examples may not have)
plugins.md031.enabled = false  # Allow code blocks without surrounding blank lines (nested examples)
extensions.front-matter.enabled = true  # Support YAML front matter
extensions.tables.enabled = true  # Enable table support
